+++ plugin-src/game-patch/include/utils.h
@@ -1,5 +1,5 @@
 #include <Common.h>
 #include "plugin_common.h"
 
-s32 Read_File(const char *File, char **Data, size_t *Size, int extra);
-s32 Write_File(const char *File, unsigned char *Data, size_t Size);
+s32 Read_File(const char *File, char **Data, u64 *Size, u32 extra);
+s32 Write_File(const char *File, unsigned char *Data, u64 Size);
